Move pie chart into a seperate component, remove redundant interface for the map
