+++ ClientApp/src/Components/StatsDashboard/index.js
@@ -10,7 +10,5 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
     for (var p in m) if (p !== "default" && !Object.prototype.hasOwnProperty.call(exports, p)) __createBinding(exports, m, p);
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-__exportStar(require("./Menu"), exports);
-__exportStar(require("./DWMapComponents"), exports);
-__exportStar(require("./StatsDashboard"), exports);
+__exportStar(require("./LayerPieChart"), exports);
 //# sourceMappingURL=index.js.map
