+++ ClientApp/src/Components/index.ts
@@ -1,2 +1,3 @@
 ﻿export * from './Menu';
 export * from './DWMapComponents';
+export * from './StatsDashboard';
