+++ ClientApp/src/Interfaces/Data/StatsDashboard/ILayerTotals.ts
@@ -1,0 +1,6 @@
+﻿export interface ILayerTotals {
+    layerId: number,
+    layerTitle: string,
+    layerColour: string,
+    uses: number
+}
