+++ ClientApp/src/Interfaces/Data/index.js
@@ -11,4 +11,5 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 __exportStar(require("./Routing"), exports);
+__exportStar(require("./StatsDashboard"), exports);
 //# sourceMappingURL=index.js.map
