+++ ClientApp/src/Interfaces/Data/index.ts
@@ -1,1 +1,2 @@
 ﻿export * from './Routing';
+export * from './StatsDashboard';
