--- ClientApp/src/Interfaces/Screens/IMap.ts
@@ -1,5 +1,0 @@
-﻿import { Layer } from 'maptalks';
-
-export interface IMap {
-    className: string
-}
