--- ClientApp/src/Interfaces/Screens/index.ts
@@ -1,1 +1,0 @@
-﻿export * from './IMap';
