+++ ClientApp/src/Interfaces/index.ts
@@ -1,4 +1,3 @@
 ﻿export * from './Layers';
-export * from './Screens';
 export * from './Components';
 export * from './Data';
