+++ RailVisualisation.csproj
@@ -42,6 +42,8 @@
     <None Remove="ClientApp\src\Components\Menu\index.ts" />
     <None Remove="ClientApp\src\Components\Menu\search.tsx" />
     <None Remove="ClientApp\src\Components\Menu\SearchItem.tsx" />
+    <None Remove="ClientApp\src\Components\StatsDashboard\index.ts" />
+    <None Remove="ClientApp\src\Components\StatsDashboard\layerPieChart.tsx" />
     <None Remove="ClientApp\src\Interfaces\Components\IMenu.ts" />
     <None Remove="ClientApp\src\Interfaces\Components\index.ts" />
     <None Remove="ClientApp\src\Interfaces\Components\Menu\ILayerMenuItem.ts" />
@@ -56,12 +58,12 @@
     <None Remove="ClientApp\src\Interfaces\Data\Routing\IRoutes.ts" />
     <None Remove="ClientApp\src\Interfaces\Data\Routing\IStep.ts" />
     <None Remove="ClientApp\src\Interfaces\Data\Routing\ITextValuePair.ts" />
+    <None Remove="ClientApp\src\Interfaces\Data\StatsDashboard\ILayerTotals.ts" />
+    <None Remove="ClientApp\src\Interfaces\Data\StatsDashboard\index.ts" />
     <None Remove="ClientApp\src\Interfaces\index.ts" />
     <None Remove="ClientApp\src\Interfaces\Layers\ILayer.ts" />
     <None Remove="ClientApp\src\Interfaces\Layers\index.ts" />
     <None Remove="ClientApp\src\Interfaces\Layers\IRoutingItem.ts" />
-    <None Remove="ClientApp\src\Interfaces\Screens\IMap.ts" />
-    <None Remove="ClientApp\src\Interfaces\Screens\index.ts" />
     <None Remove="ClientApp\src\Layers\Routing\DirectionsItem.tsx" />
     <None Remove="ClientApp\src\Layers\Routing\index.ts" />
     <None Remove="ClientApp\src\Layers\Routing\Routing.tsx" />
